コンパイル検証で見つかった問題を修正 (全スケッチ ビルド成功)
- Nano Every は D1=RX1(PC5)/D0=TX1(PC4): GPS 配線のドキュメントを修正
- INVEN 設定をコアのマクロ経由に変更 (ATMEGA328 レジスタエミュレーション有効でもビルド可)
- Wire.requestFrom のオーバーロード曖昧さを size_t キャストで解消
- 統合FW: Flash 46% / RAM 16% (arduino-cli, arduino:megaavr:nona4809 で確認)

+++ firmware/compass/mag.cpp
@@ -34,7 +34,7 @@ static bool readRegs(uint8_t reg, uint8_t *buf, uint8_t len) {
   Wire.beginTransmission(ADDR);
   Wire.write(reg);
   if (Wire.endTransmission(false) != 0) return false;
-  if (Wire.requestFrom(ADDR, len) != len) return false;
+  if (Wire.requestFrom(ADDR, (size_t)len) != len) return false;
   for (uint8_t i = 0; i < len; i++) buf[i] = Wire.read();
   return true;
 }
